pyproject.toml
[project]
name = "juntos"
version = "0.1.0"
description = "A global organizational structure centered around Ben Franklin's juntos concept"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1",
    "flask-sqlalchemy>=3.1",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
    "authlib>=1.3.2",
    "requests>=2.32",
    "python-dotenv>=1.0",
]

[project.scripts]
juntos = "juntos.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/juntos"]
